requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
yt-dlp
Pillow
requests
feedparser
numpy